Unindent subtest diag header.
Experiment in reaponse to #264. The change itself is simple enough, removing
`' '` from line 6338 of `sql/pgtap.sql.in`. But since pgTAP itself depends on
`make installcheck`, you can see all the other test `*.out` files that would
also need to be changed for tests to continue passing.

If many other projects depend on `make installcheck` and use xUnit-style
tests, they, too, would need to change, which would be a PITA -- especially if
the new version of pgTAP could be run against older releases, which would
suddenly start failing for no good reason. Or, just as bad, a new version is
released that could have tests run against an older version of pgTAP.

So pushing for consideration and discusssion, as the indentation as-is, while
the same as for Perl's Test::More, is less legible, especially when subtests
fail.